Task-2/requirements.txt
tensorflow>=2.5
numpy
pillow
tk
gradio
scikit-learn